+++ PDataIO.h
@@ -18,8 +18,8 @@
 - (int) getPasvFd:(PSession*)session;
 - (int) getPortFd:(PSession*)session;
 - (BOOL) postMarkConnect:(PSession*)session;
-- (unsigned long) sendData:(PSession*)session data:(const void*)data dataSize:(unsigned long)dataSize;
-- (unsigned long) sendFile:(PSession*)session fileHandle:(NSFileHandle*)fileHandle fileSize:(unsigned long)fileSize;
-- (unsigned long) recvFile:(PSession*)session fileHandle:(NSFileHandle*)fileHandle;
+- (int) sendData:(PSession*)session data:(const void*)data dataSize:(unsigned long)dataSize;
+- (int) sendFile:(PSession*)session fileHandle:(NSFileHandle*)fileHandle fileSize:(unsigned long)fileSize;
+- (int) recvFile:(PSession*)session fileHandle:(NSFileHandle*)fileHandle;
 
 @end
